% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_match.R
\name{ps_match}
\alias{ps_match}
\title{Propensity score matching with fixed matching ratio.}
\usage{
ps_match(input_data, match_args)
}
\arguments{
\item{input_data}{Data with patient id, treatment status, covariates, and 
outcomes, as returned by plasmode().}

\item{match_args}{List of arguments for matching algorithm as specified in 
match_params.R.}
}
\value{
Matched data in identical format to the initial data with id, 
  covariate, treatment, and outcome columns.
}
\description{
Propensity score matching with fixed matching ratio for use in 
  TBSS with Bernoulli probability model. Arguments are specified in the file
  match_params.R
}
